introduce a more flexible way to define affectedBy relationships of keyword
instead of define a affectedBy relationship with list of strings we us
a list of new introduces object AffectedBy with a name (the old
string) and a new AffectByType. The AffectByType can be
Extends this only shows that the keyword stays not alone and the
validation result can be altered. Or Replace with how the names say
replace the affected keyword with the affectedBy keyword. this allow
to change the validation result in both directions und a more flexible
way to do it.