introduce a more flexible way to define affects relationships of keyword
instead of define a affects relationship with list of strings we us
a list of new introduces object Affects with a name (the old
string) and a new AbsenceStrategy. The AbsenceStrategy can be use to
define what should be happens if the keyword is missing.
At default is exists two:
Provide a default value -> useful if the affected works without the affectedBy one
Replace -> useful if the affected doesn't works without the affectedBy one