remove usage of record for VocabularyDefinition
VocabularyDefinition is a interface now, the default implemention move
into LazyVocabularyDefinition which no longer use the serviceloader
directly. the vocabulary avaivable for a definitions will provided via a
supplier. ServiceLoaderLazyVocabulariesSupplier can be use to use a
ServiceLoader to find a requested vocabulary.
VocabularyDefinition doesn't provide a method to retrieve the id instead
it defines a method to ask if it has a given id (hasId("") instead of
id().equals("")).